2.0

+++ api/src/v1.0/controller/api.js
@@ -42,6 +42,10 @@ export default class extends Base {
 
     delete data[pk]
 
+    if (!await this.memberOf(data.project_id)) {
+      return this.fail('NOT_MEMBER')
+    }
+
     if (think.isEmpty(data)) {
       return this.fail('data is empty')
     }
@@ -49,6 +53,16 @@ export default class extends Base {
     data.user_id = user.id
 
     data.id = await this.modelInstance.add(data)
+
+    if (data.id) {
+      let projectModelInstance = this.model('project')
+      let project = await projectModelInstance.where({ id: data.project_id }).find()
+
+      await this.createLogger({
+        description: `创建了项目【${project.name}】接口【${data.name}】`,
+        project_id: data.project_id
+      })
+    }
 
     return this.success(data)
   }
@@ -59,7 +73,23 @@ export default class extends Base {
     }
 
     let pk = await this.modelInstance.getPk()
+    let api = await this.modelInstance.where({ [pk]: this.id }).find()
+
+    if (!await this.memberOf(api.project_id)) {
+      return this.fail('NOT_MEMBER')
+    }
+
     let rows = await this.modelInstance.where({ [pk]: this.id }).delete()
+
+    if (rows > 0) {
+      let projectModelInstance = this.model('project')
+      let project = await projectModelInstance.where({ id: api.project_id }).find()
+
+      await this.createLogger({
+        description: `删除了项目【${project.name}】接口【${api.name}】`,
+        project_id: api.project_id
+      })
+    }
 
     return this.success({ affectedRows: rows })
   }
@@ -72,6 +102,10 @@ export default class extends Base {
     let pk = await this.modelInstance.getPk()
     let data = this.post()
 
+    if (!await this.memberOf(data.project_id)) {
+      return this.fail('NOT_MEMBER')
+    }
+
     delete data[pk]
     delete data.user
     delete data.response
@@ -83,6 +117,16 @@ export default class extends Base {
     }
 
     let rows = await this.modelInstance.where({ [pk]: this.id }).update(data)
+
+    if (rows) {
+      let projectModelInstance = this.model('project')
+      let project = await projectModelInstance.where({ id: data.project_id }).find()
+
+      await this.createLogger({
+        description: `更新了项目【${project.name}】接口【${data.name}】信息`,
+        project_id: data.project_id
+      })
+    }
 
     return this.success({ affectedRows: rows })
   }
