+++ api/src/v1.0/controller/module.js
@@ -42,6 +42,10 @@ export default class extends Base {
 
     delete data[pk]
 
+    if (!await this.memberOf(data.project_id)) {
+      return this.fail('NOT_MEMBER')
+    }
+
     if (think.isEmpty(data)) {
       return this.fail('data is empty')
     }
@@ -50,7 +54,15 @@ export default class extends Base {
 
     data.id = await this.modelInstance.add(data)
 
-    await this.createLogger({ title: `创建模块【${data.name}】`, project_id: data.project_id })
+    if (data.id) {
+      let projectModelInstance = this.model('project')
+      let project = await projectModelInstance.where({ id: data.project_id }).find()
+
+      await this.createLogger({
+        description: `创建了项目【${project.name}】接口分组【${data.name}】`,
+        project_id: data.project_id
+      })
+    }
 
     return this.success(data)
   }
@@ -61,7 +73,37 @@ export default class extends Base {
     }
 
     let pk = await this.modelInstance.getPk()
-    let rows = await this.modelInstance.where({ [pk]: this.id }).delete()
+    let module = await this.modelInstance.where({ [pk]: this.id }).find()
+    let projectModelInstance = this.model('project')
+    let project = await projectModelInstance.where({ id: module.project_id }).find()
+
+    if (!await this.ownerOf(project)) {
+      return this.fail('NOT_OWNER')
+    }
+
+    let rows = await this.modelInstance.transaction(async () => {
+      let apiModelInstance = this.model('api')
+      let responseModelInstance = this.model('response')
+      let paramModelInstance = this.model('param')
+      let fieldModelInstance = this.model('field')
+      let apis = await apiModelInstance.where({ module_id: this.id }).select()
+
+      for (let api in apis) {
+        await paramModelInstance.where({ api_id: api.id }).delete()
+        await fieldModelInstance.where({ api_id: api.id }).delete()
+        await responseModelInstance.where({ api_id: api.id }).delete()
+        await apiModelInstance.where({ id: api.id }).delete()
+      }
+
+      return await this.modelInstance.where({ [pk]: this.id }).delete()
+    })
+
+    if (rows > 0) {
+      await this.createLogger({
+        description: `删除了项目【${project.name}】接口分组【${module.name}】`,
+        project_id: module.project_id
+      })
+    }
 
     return this.success({ affectedRows: rows })
   }
@@ -74,6 +116,10 @@ export default class extends Base {
     let pk = await this.modelInstance.getPk()
     let data = this.post()
 
+    if (!await this.memberOf(data.project_id)) {
+      return this.fail('NOT_MEMBER')
+    }
+
     delete data[pk]
     delete data.user
     delete data.created_at
@@ -84,7 +130,17 @@ export default class extends Base {
     }
 
     let rows = await this.modelInstance.where({ [pk]: this.id }).update(data)
-    await this.createLogger({ title: `更新模块【${data.name}】`, project_id: data.project_id })
+
+    if (rows) {
+      let projectModelInstance = this.model('project')
+      let project = await projectModelInstance.where({ id: data.project_id }).find()
+
+      await this.createLogger({
+        description: `更新了项目【${project.name}】接口分组【${data.name}】信息`,
+        project_id: data.project_id
+      })
+    }
+
     return this.success({ affectedRows: rows })
   }
 }
