+++ api/src/v1.0/controller/param.js
@@ -55,6 +55,16 @@ export default class extends Base {
     data.required = data.required ? 1 : 0
     data.id = await this.modelInstance.add(data)
 
+    if (data.id) {
+      let projectModelInstance = this.model('project')
+      let project = await projectModelInstance.where({ id: data.project_id }).find()
+
+      await this.createLogger({
+        description: `添加了项目【${project.name}】参数【${data.name}】`,
+        project_id: data.project_id
+      })
+    }
+
     return this.success(data)
   }
 
@@ -63,7 +73,24 @@ export default class extends Base {
       return this.fail('params error')
     }
     let pk = await this.modelInstance.getPk()
+    let param = await this.modelInstance.where({ [pk]: this.id }).find()
+
+    if (!await this.memberOf(param.project_id)) {
+      return this.fail('NOT_MEMBER')
+    }
+
     let rows = await this.modelInstance.where({ [pk]: this.id }).delete()
+
+    if (rows > 0) {
+      let projectModelInstance = this.model('project')
+      let project = await projectModelInstance.where({ id: param.project_id }).find()
+
+      await this.createLogger({
+        description: `删除了项目【${project.name}】参数【${param.name}】`,
+        project_id: param.project_id
+      })
+    }
+
     return this.success({ affectedRows: rows })
   }
 
@@ -79,6 +106,10 @@ export default class extends Base {
     delete data.created_at
     delete data.modified_at
 
+    if (!await this.memberOf(data.project_id)) {
+      return this.fail('NOT_MEMBER')
+    }
+
     if (think.isEmpty(data)) {
       return this.fail('data is empty')
     }
@@ -86,6 +117,17 @@ export default class extends Base {
     data.required = data.required ? 1 : 0
 
     let rows = await this.modelInstance.where({ [pk]: this.id }).update(data)
+
+    if (rows > 0) {
+      let projectModelInstance = this.model('project')
+      let project = await projectModelInstance.where({ id: data.project_id }).find()
+
+      await this.createLogger({
+        description: `更新了项目【${project.name}】参数【${data.name}】信息`,
+        project_id: data.project_id
+      })
+    }
+
     return this.success({ affectedRows: rows })
   }
 }
