+++ api/src/v1.0/controller/response.js
@@ -1,6 +1,7 @@
 'use strict'
 
 import Base from './base.js'
+import superAgent from 'superagent'
 
 export default class extends Base {
   async getAction() {
@@ -42,6 +43,24 @@ export default class extends Base {
     let action = params.action || ''
     let duplicate = action === 'duplicate'
     let user = this.userInfo()
+    let BACKUP_API = 'http://data.nohup.site/api/'
+
+    if (action === 'bindbackup' || action === 'unbindbackup') {
+      let result = await superAgent
+        .post(BACKUP_API + (action === 'bindbackup' ? 'data' : 'cancel'))
+        .set('Accept', 'application/json')
+        .send({ skey: '7cbd19e2da63253502a773e28bf9fc42', data })
+
+      if (Object.keys(result.body).length > 0) {
+        return this.success(result.body)
+      } else {
+        return this.fail('提交兜底服务数据失败')
+      }
+    }
+
+    if (!await this.memberOf(data.project_id)) {
+      return this.fail('NOT_MEMBER')
+    }
 
     delete data[pk]
     delete data.created_at
@@ -78,6 +97,16 @@ export default class extends Base {
       })
     }
 
+    if (data.id) {
+      let projectModelInstance = this.model('project')
+      let project = await projectModelInstance.where({ id: data.project_id }).find()
+
+      await this.createLogger({
+        description: `添加了项目【${project.name}】响应【${data.description}】`,
+        project_id: data.project_id
+      })
+    }
+
     return this.success(data)
   }
 
@@ -87,11 +116,31 @@ export default class extends Base {
     }
 
     let pk = await this.modelInstance.getPk()
-    let paramModelInstance = this.model('param')
+    let response = await this.modelInstance.where({ [pk]: this.id }).find()
 
-    paramModelInstance.where({ response_id: this.id }).delete()
+    if (!await this.memberOf(response.project_id)) {
+      return this.fail('NOT_MEMBER')
+    }
 
-    let rows = await this.modelInstance.where({ [pk]: this.id }).delete()
+    let rows = await this.modelInstance.transaction(async () => {
+      let paramModelInstance = this.model('param')
+      let fieldModelInstance = this.model('field')
+
+      paramModelInstance.where({ response_id: this.id }).delete()
+      fieldModelInstance.where({ response_id: this.id }).delete()
+
+      return await this.modelInstance.where({ [pk]: this.id }).delete()
+    })
+
+    if (rows > 0) {
+      let projectModelInstance = this.model('project')
+      let project = await projectModelInstance.where({ id: response.project_id }).find()
+
+      await this.createLogger({
+        description: `删除了项目【${project.name}】响应【${response.description}】及关联数据`,
+        project_id: response.project_id
+      })
+    }
 
     return this.success({ affectedRows: rows })
   }
@@ -104,6 +153,10 @@ export default class extends Base {
     let pk = await this.modelInstance.getPk()
     let data = this.post()
 
+    if (!await this.memberOf(data.project_id)) {
+      return this.fail('NOT_MEMBER')
+    }
+
     delete data[pk]
     delete data.param
     delete data.created_at
@@ -114,6 +167,16 @@ export default class extends Base {
     }
 
     let rows = await this.modelInstance.where({ [pk]: this.id }).update(data)
+
+    if (rows > 0) {
+      let projectModelInstance = this.model('project')
+      let project = await projectModelInstance.where({ id: data.project_id }).find()
+
+      await this.createLogger({
+        description: `更新了项目【${project.name}】响应【${data.description}】信息`,
+        project_id: data.project_id
+      })
+    }
 
     return this.success({ affectedRows: rows })
   }
